+++ src/graphql/treatment_type/treatment_type.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { CreateTreatmentTypeInput } from './dto/create-treatment_type.input';
 import { UpdateTreatmentTypeInput } from './dto/update-treatment_type.input';
 import { PrismaService } from 'src/prisma/prisma.service';
+import { GraphQLError } from 'graphql';
 
 @Injectable()
 export class TreatmentTypeService {
@@ -17,27 +18,50 @@ export class TreatmentTypeService {
   }
 
   async findOne(id: number) {
-    return await this.prisma.treatmentType.findUniqueOrThrow({
-      where:{id:id},
-    }) ;
+    const treatmentType = await this.prisma.treatmentType.findUnique({
+      where: {id: id},
+    }) 
+    if (!treatmentType) {
+      throw new GraphQLError('treatmentType not found', {
+        extensions: {
+          code: 404,
+        },
+      });
+    }
+    return  treatmentType;
   }
 
-   async update(id: number, updateTreatmentTypeInput: UpdateTreatmentTypeInput) {
-   await this.prisma.treatmentType.findUniqueOrThrow({
-      where:{ id :id},
-    });
+  async update(id: number, updateTreatmentTypeInput: UpdateTreatmentTypeInput) {
+    const treatmentType = await this.prisma.treatmentType.findUnique({
+      where: {id: id},
+    }) 
+
+    if (!treatmentType) {
+      throw new GraphQLError('treatmentType not found', {
+        extensions: {
+          code: 404,
+        },
+      });
+    }
     return await this.prisma.treatmentType.update({
-      where:{id : id},
-      data:{name : updateTreatmentTypeInput.name},
+      where:{id:id},
+      data:{name:updateTreatmentTypeInput.name}
     });
   }
 
   async remove(id: number) {
-   await this.prisma.treatmentType.findUniqueOrThrow({
-      where:{ id :id},
-    });
-    return this.prisma.treatmentType.delete({
-      where:{id : id},
+    const treatmentType = await this.prisma.treatmentType.findUnique({
+      where: {id: id},
+    }) 
+    if (!treatmentType) {
+      throw new GraphQLError('treatmentType not found', {
+        extensions: {
+          code: 404,
+        },
+      });
+    }
+    return await this.prisma.treatmentType.delete({
+      where:{id:id},
     });
   }
 }
